requirements.txt
kernel
playwright
python-dotenv